feat: add confirmation dialog for user deletion in ListFragment

+++ app/src/main/java/com/example/myapplication/ui/ListFragment.kt
@@ -4,6 +4,8 @@ import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.Toast
+import androidx.appcompat.app.AlertDialog
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.Observer
@@ -48,7 +50,18 @@ class ListFragment : Fragment() {
     }
 
     private fun onDeleteUser(user: User) {
-        userViewModel.deleteUser(user)
+
+        val builder = AlertDialog.Builder(requireContext())
+
+        builder.setPositiveButton("YES") { _, _ ->
+            userViewModel.deleteUser(user)
+            Toast.makeText(requireContext(), "Successfully removed", Toast.LENGTH_SHORT).show()
+        }
+
+        builder.setNegativeButton("NO") { _, _ -> }
+        builder.setTitle("Delete ${user.firstName}")
+        builder.setMessage("Are you sure you want to delete ${user.firstName}?")
+        builder.create().show()
     }
